Adicionando função de acionamento do buzzer
Criando arquivo .c e .h para integração dos arquivos na compilação do projeto isolando o método de acionamento do buzzer facilitando manutenção de código
